ReportVordergrund und Hintergrund in Reporter verschoben
